LinearSparseSolveAndInverseTest: Clean up warnings
Cleaning up warnings from variable type assignmments.

+++ SurgSim/Math/UnitTests/LinearSparseSolveAndInverseTests.cpp
@@ -43,11 +43,12 @@ private:
 	void initializeSparseMatrix(SparseMatrix* m)
 	{
 		m->resize(static_cast<int>(size), static_cast<int>(size));
-		for (int row = 0; row < size; row++)
+		for (size_t row = 0; row < size; row++)
 		{
-			for (int col = 0; col < size; col++)
+			for (size_t col = 0; col < size; col++)
 			{
-				(*m).insert(row, col) = fmod((10.3 * cos(static_cast<double>(row * col)) + 3.24), 10.0);
+				(*m).insert(static_cast<int>(row), static_cast<int>(col)) =
+					fmod((10.3 * cos(static_cast<double>(row * col)) + 3.24), 10.0);
 			}
 		}
 		m->makeCompressed();
